frontend/src/app/view-submitted-text-message-template/[templateId]/page.tsx
'use server';

import { PageProps } from 'nhs-notify-web-template-management-utils';
import { getTemplate } from '@utils/form-actions';
import { redirect, RedirectType } from 'next/navigation';
import { ViewSMSTemplate } from '@molecules/ViewSMSTemplate/ViewSMSTemplate';
import { validateSubmittedSMSTemplate } from '@utils/validate-template';

const ViewSubmittedSMSTemplatePage = async ({
  params: { templateId },
}: PageProps) => {
  const template = await getTemplate(templateId);

  const validatedTemplate = validateSubmittedSMSTemplate(template);

  if (!validatedTemplate) {
    redirect('/invalid-template', RedirectType.replace);
  }

  return <ViewSMSTemplate initialState={validatedTemplate} />;
};

export default ViewSubmittedSMSTemplatePage;
